chore: update workspace config
