+++ .changeset/config.json
@@ -2,7 +2,7 @@
   "$schema": "https://unpkg.com/@changesets/config@3.1.1/schema.json",
   "changelog": [
     "@svitejs/changesets-changelog-github-compact",
-    { "repo": "TanStack/optimistic" }
+    { "repo": "TanStack/db" }
   ],
   "commit": false,
   "access": "public",
