+++ .npmrc
@@ -1,3 +1,1 @@
-link-workspace-packages=true
-prefer-workspace-packages=true
 provenance=true
