+++ .nvmrc
@@ -1,1 +1,1 @@
-22.13.1
+24.8.0
