+++ packages/electric-db-collection/tests/electric.test-d.ts
@@ -200,7 +200,7 @@ describe(`Electric collection type resolution tests`, () => {
       query: (q) =>
         q
           .from({ user: usersCollection })
-          .where(({ user }) => eq(user.active, true) && gt(user.age, 18)) // eslint-disable-line @typescript-eslint/no-unnecessary-condition
+          .where(({ user }) => eq(user.active, true) && gt(user.age, 18))
           .select(({ user }) => ({
             id: user.id,
             name: user.name,
