+++ packages/query-db-collection/src/query.ts
@@ -338,7 +338,6 @@ export function queryCollectionOptions(
     throw new QueryClientRequiredError()
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
   if (!getKey) {
     throw new GetKeyRequiredError()
   }
