+++ scripts/verify-links.ts
@@ -1,6 +1,6 @@
 import { existsSync, readFileSync, statSync } from "node:fs"
 import path, { resolve } from "node:path"
-import fg from "fast-glob"
+import { glob } from "tinyglobby"
 // @ts-ignore - markdown-link-extractor is not typed
 import markdownLinkExtractor from "markdown-link-extractor"
 
@@ -87,7 +87,7 @@ function fileExistsForLink(
 
 async function findMarkdownLinks() {
   // Find all markdown files in docs directory
-  const markdownFiles = await fg(`docs/**/*.md`, {
+  const markdownFiles = await glob(`docs/**/*.md`, {
     ignore: [`**/node_modules/**`],
   })
